MapRoulette refresh: Continue on 504
Sometimes -- presumably when the refresh takes too long -- the
MapRoulette API would respond with an HTTP error 504 "Gateway Timeout".
In the web frontend one can see, though, that the rebuilding of the
challenge is under way. So, assuming that the refresh is triggered as
background process, continue anyway when receiving a 504 error.