Building area fraction, split landuse by highways
Calculate the areas of landuse and building polygons. Select landuse
polygons which are of a minimum size and below a threashold of building
coverage. The threshold is selected to be 5 % for now, because that
seemed plausible and to generate not too many false positives.

To make the MapRoulette tasks easier to complete, split the affected
landuse areas into smaller areas as divided by streets. It was also
feedback of the mappers of the previous round that too large tasks are
demotivating.

As there can be very small or thin polygons between parallel highways,
e.g. for dual highways, merge these "sliver" polygons again. These would
generate non-sensical false-positive tasks.