Split tasks by German states
Putting all tasks into one challenge in MapRoulette is just too big and
makes it sluggish.

Split the tasks by German states, similar to how it was already done in
v1.

Berlin, Bremen, Hamburg and Saarland are so small and have so few tasks
that they are aggregated into one challenge.