Improve sliver polygon merging
Instead of requiring the neighbor of a sliver-polygon to be non-sliver,
get the neighbor with the longest shared border. Then find the
transitive closure of polygons to merge. Merge them all in one step,
instead of the previous while-loop.

Increase the minumum inscribed circle radius to 50 m, as the feedback
was that the polygons are too small. This leads to more small polygons
being merged.