Subtract "unwanted" areas from residential
In many places, the landuse=residential areas are big and are overlapped
by other land uses such as parks, cemeteries, schools and so on. As
these don't contain residential buildings they would be flagged as
tasks.

Instead, subtract non-residential landuses from the residential areas at
the beginning of the processing.

Reduce they types of highways used for splitting. In some places there
are footways or cycleways mapped on both sides of a (bigger) road. The
area inbetween is large enough to be counted as residential without
buildings. In order to avoid these false positives, only use highways
"for cars" for splitting the landuse areas.

The merging of sliver polygons was very slow. Speed it up a lot, by
creating a necessar index earlier. Also add progress logs.